Change static variable starting field to `schema`
This change is introduced to ensure we know which spec we're
referencing. We are still only able to reference the `spec`
and the `metadata` of the schema.

This change also affects how we reference schema variables
when compiling `includeWhen` expressions

changing from
```yaml
name: ${spec.name}
```

to
```yaml
name: ${schema.spec.name}
```

more examples in files changed